[MOV] sale_quotation_builder: split sale_order_template.py
